Refactor import paths from interface to interfaces for consistency; add interfaces file for FooterMultiColumn components
