+++ components/organisms/FooterMultiColumn/interfaces.ts
